+++ microgradx/optim/__init__.py
@@ -3,19 +3,22 @@ from microgradx.optim.sgd import SGD
 from microgradx.optim.adam import Adam
 from microgradx.optim.adamw import AdamW
 from microgradx.optim.lion import Lion
+from microgradx.optim.radam import RAdam
 from microgradx.optim.lr_scheduler import (
     LambdaLR,
     StepLR,
     MultiStepLR,
     ExponentialLR,
     CosineAnnealingLR,
+    CosineAnnealingWarmRestarts,
     LinearWarmup,
     OneCycleLR,
     ReduceLROnPlateau,
 )
 
-__all__ = ["Optimizer", "SGD", "Adam", "AdamW", "Lion",
+__all__ = ["Optimizer", "SGD", "Adam", "AdamW", "Lion", "RAdam",
            "clip_grad_norm_", "clip_grad_value_",
            "LambdaLR", "StepLR", "MultiStepLR", "ExponentialLR",
-           "CosineAnnealingLR", "LinearWarmup", "OneCycleLR",
+           "CosineAnnealingLR", "CosineAnnealingWarmRestarts",
+           "LinearWarmup", "OneCycleLR",
            "ReduceLROnPlateau"]